+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3
+- Improve aesthetics of error messages
+- Warnings no longer block output from showing up
+- Add clipboard support
+
 ## 0.1.2
 - Add separate message dialogs for info, warning, and errors